Attempt at removing blocking impl: mayhem
