+++ htsget-search/Cargo.toml
@@ -17,13 +17,13 @@ htsget-id-resolver = { path = "../htsget-id-resolver"}
 
 noodles = { version = "0.18.0", features = ["bgzf", "bam", "bcf", "cram", "csi", "sam", "tabix", "vcf"] }
 
-# noodles-bgzf = { version = "0.8.0", features = ['async'] }
-# noodles-bam = { version = "0.14.0", features = ['async'] }
-# noodles-bcf = { version = "0.12.0", features = ['async'] }
-# noodles-cram = { version = "0.11.0", features = ['async'] }
-# noodles-csi = { version = "0.4.3", features = ['async'] }
-# noodles-tabix = { version = "0.7.3", features = ['async'] }
-# noodles-vcf = { version = "0.13.0", features = ['async'] }
+noodles-bgzf = { version = "0.8.0", features = ['async'] }
+noodles-bam = { version = "0.14.0", features = ['async'] }
+noodles-bcf = { version = "0.12.0", features = ['async'] }
+noodles-cram = { version = "0.11.0", features = ['async'] }
+noodles-csi = { version = "0.4.3", features = ['async'] }
+noodles-tabix = { version = "0.7.3", features = ['async'] }
+noodles-vcf = { version = "0.13.0", features = ['async'] }
 
 [dev-dependencies]
 tempfile = "3.2.0"
