+++ htsget-search/src/lib.rs
@@ -1,4 +1,2 @@
-//! Minimal Viable Product (MVP) to prove that an implementation of the [HtsGet specification](https://samtools.github.io/hts-specs/htsget.html) is worth using Rust.
-//!
 pub mod htsget;
 pub mod storage;
