+++ htsget-search/src/storage/mod.rs
@@ -4,15 +4,11 @@ use std::cmp::Ordering;
 
 use thiserror::Error;
 
-#[cfg(feature = "async")]
+pub mod async_storage;
 pub use async_storage::*;
 
 use crate::htsget::Class;
 
-#[cfg(feature = "async")]
-pub mod async_storage;
-pub mod blocking;
-#[cfg(feature = "async")]
 pub mod local;
 
 type Result<T> = core::result::Result<T, StorageError>;
